requirements.txt
streamlit
pandas
python-dotenv
openai
langchain
langchain_community